requirements.txt
kivy
pillow
tqdm
numpy
scikit-learn
torch
transformers
pynput
PyQt5